[IMP] crm_claim_type changed to crm_claim_rma_config module
