+++ crm_claim_rma/__openerp__.py
@@ -29,9 +29,8 @@
     'depends': ['sale',
                 'sales_team',
                 'stock',
-                'crm_claim',
-                'product_warranty',
                 'crm_claim_rma_config',
+                'product_warranty',
                 'crm_rma_location_rma',
                 ],
     'data': ['wizard/claim_make_picking_view.xml',
