+++ crm_claim_rma/crm_claim_rma_data.xml
@@ -89,22 +89,5 @@
             <field name="object_id" search="[('model','=','crm.claim')]" model="ir.model"/>
         </record>
 
-
-    <!-- Claims Type -->
-    <record id="crm_claim_type_supplier" model="crm.claim.type">
-        <field name="name">Supplier</field>
-        <field name="active">1</field>
-    </record>
-
-    <record id="crm_claim_type_customer" model="crm.claim.type">
-        <field name="name">Customer</field>
-        <field name="active">1</field>
-    </record>
-
-    <record id="crm_claim_type_other" model="crm.claim.type">
-        <field name="name">Other</field>
-        <field name="active">1</field>
-    </record>
-
     </data>
 </openerp>
