+++ crm_claim_rma_config/__init__.py
@@ -18,3 +18,5 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
+
+from . import crm_claim_rma
