+++ crm_claim_rma_config/crm_claim_rma_data.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+<openerp>
+    <data noupdate="1">
+
+    <!-- Claims Type -->
+    <record id="crm_claim_type_supplier" model="crm.claim.type">
+        <field name="name">Supplier</field>
+        <field name="active">1</field>
+    </record>
+
+    <record id="crm_claim_type_customer" model="crm.claim.type">
+        <field name="name">Customer</field>
+        <field name="active">1</field>
+    </record>
+
+    <record id="crm_claim_type_other" model="crm.claim.type">
+        <field name="name">Other</field>
+        <field name="active">1</field>
+    </record>
+
+    </data>
+</openerp>
