+++ crm_claim_rma_config/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
+"access_crm_claim_type_user","crm.claim.type.user","model_crm_claim_type","base.group_sale_salesman","True","True","True","True"
+
